dùng trigger tạo id_laptop

+++ Controllers/HeartController.cs
@@ -34,7 +34,7 @@ namespace LaptopShop.Controllers
                         Id = temp.IdLaptop,
                         Name = temp.TenLapTop,
                         Hinh = temp.HinhAnh ?? "",
-                        Price = temp.GiaBan ?? 0
+                        Price = (double)(temp.GiaBan ?? 0)
                     };
                     danhsach.Add(item);
                 }
